Adiciona SweetAlert2

+++ src/app/escola/escola.component.ts
@@ -2,12 +2,15 @@ import { EscolaService } from './../escola.service';
 import { Component, OnInit } from '@angular/core';
 import { Escola } from '../escola';
 
+declare let swal: any;
+
 @Component({
   selector: 'app-escola',
   templateUrl: './escola.component.html',
   styleUrls: ['./escola.component.css']
 })
 export class EscolaComponent implements OnInit {
+  
   escolas: Array<Escola>;
   selectedEscola: Escola;
 
@@ -27,8 +30,21 @@ export class EscolaComponent implements OnInit {
   }
 
   delete(escola: Escola): void{
-    this.escolas = this.escolas.filter(x => x !== escola);
-    this.escolaService.deleteEscola(escola).subscribe();
+    swal({
+      title: "",
+      text: "Deseja excluir a escola? Isso excluirá também todas as turmas.",
+      type: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Excluir",
+      confirmButtonClass: 'btn btn-success',
+      cancelButtonText: "Cancelar",
+      cancelButtonClass: 'btn btn-danger'
+    }).then((result) => {
+      if(result.value){
+        this.escolas = this.escolas.filter(x => x !== escola);
+        this.escolaService.deleteEscola(escola).subscribe();
+      }
+    })
   }
 
 }
